core/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "AI Agent module for core functionalities"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.43",
    "botocore>=1.42.43",
    "pydantic-ai-slim[bedrock,logfire]>=1.56.0",
]

[dependency-groups]
dev = [
    "boto3-stubs>=1.42.43",
    "botocore-stubs>=1.42.41",
]

[tool.uv.sources]
shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
